# Optimize regex star operation