@pracucci 2023-03-08
> Because .+ is what Grafana sends as All value for templating variables.

Right, but the customer is using `.*` so it's important we optimise `.*` too for the case I'm investigating.