@pracucci 2023-03-08
> Hopefully IndexByte uses the magic SIMD instructions to make it quicker, but unfortunately we still have to call it on each string separately.

`strings.ContainsRune()` was already calling `strings.IndexByte()` behind the hood, so no performance difference is expected from this change (see more details in the PR description).