@duricanikolic 2023-03-13
Hi @pracucci @colega,
Shouldn’t we document that this PR breaks deep equality of `label.Matcher`’s? Maybe in `CHANGELOG.md` or in the PR description? 
I realized that while I was [upgrading mimir and mimir-prometheus](https://github.com/grafana/transmog/pull/1722)  in `transmog` and I had to fix some failing tests.